Let collections set custom feed titles
For collection feeds, you may want more control over the name of your
feed. Previously, collection feeds had titles like this:

> Site Title | Collection

This is an excellent fallback. But in some circumstances it is not
acceptable:

> Site Title | Secondary_blog_about_mysql

The site administrator may instead want a cleaner title, like "Site
Title | Secondary Blog About MySQL" or even "Blog about MySQL".

Because Jekyll collections can have arbitrary settings, we can use them
to set things like `title`s:

    collections:
      secondary_blog_about_mysql:
        title: "Blog about MySQL"
        output: true

Because this is a part of the global `collections` configuration, we can
use this new `title` attribute elsewhere on the site -- not just in
relation to the feed.