MetaTag: when encoding for XML special characters, handle non-string objects
The jekyll-github-metadata plugin, for example, sets `site.title` and
`site.name` to `Jekyll::GitHubMetadata::Value` objects which respond to
`#to_s` and `#to_liquid`, but NOT `#encode`. Therefore, we should cast
to a string if `#encode` is not yet available.

The reason we cast to a string is because the previous code was
"#{k}=#{v}", which was implicitly casting to a string for `v`.