Feed template should always have a title assigned
This commit makes clearer that the feed's <title> is always assigned to
*something*.

- For the default `posts` collection, it will always be the site title
  or site name.
- For other collections it is set to the `page.title`, which falls back
  to the `site.name`.

You can see how a non-posts collection's  `page.title` is generated in
the `generator.rb` file. If the the collection has a `title` attribute
in the YAML, we can use that. If not, we fall back to a default format:

    My Site Title | Collection_name

Before this commit, we used the `@site["name"]` attribute to generate
the default title format. But I think this is dangerous as I myself
don't have a `@site["name"]` value set... only a `@site["title"]`.