Add "html_summaries" option to allow summaries to output as full HTML
This allows posts with autogenerated excerpts (that may have images or HTML in them) to keep that HTML intact in the summary portion of the Atom feed. The default remains as having the HTML be stripped and the whitespace normalized.

This has some benefits, because the default value for a post excerpt is for it to be the entire first paragraph of each post. It's not unheard of for it to have plenty of HTML formatting and other tags.

In [my site][1], for example, post excerpts often have images, and I'd very much prefer keeping those images intact in the feed, even if using `excerpt_only` feed generation.

Note that this passes the W3C validator with no warnings, since feed.xml already uses `type="html"` in its `<summary>` tags.

[1]: https://www.kronopath.com/blog/